fix: resolve CI binary generation and Windows filesystem compatibility
This commit fixes two critical issues:

1. CI Binary Generation Issue:
- Fixed cross-compilation where all binaries had identical SHA256 hashes
- Updated Dockerfile.build to properly handle build args and binary paths
- Fixed binary extraction to use correct paths (/go/bin/{GOOS}_{GOARCH}/cm)
- Added linux/arm64 support to CI workflow
- Improved build system architecture with separate build/runtime Dockerfiles

2. Windows Filesystem Compatibility:
- Added cross-platform build tags for FileLock function
- Created pkg/fs/fs_unix.go for Unix systems (uses syscall.Flock)
- Created pkg/fs/fs_windows.go for Windows systems (no-op implementation)
- Removed platform-specific code from main fs.go file
- Windows builds now work correctly without syscall.Flock errors

3. Build System Improvements:
- Split Dockerfiles into build and runtime stages
- Made build system configuration-driven using GoImageInfo map
- Added proper error handling for unsupported build platforms
- Improved platform detection and base image selection
- Removed redundant linux/arm64/v8 platform definition

4. CI Enhancements:
- Added linux/arm64 build and release job
- All platforms now generate correct binary formats:
  - Linux AMD64: ELF binary
  - Linux ARM64: ELF binary
  - Darwin ARM64: Mach-O binary
  - Windows AMD64: PE executable (.exe)

This fix ensures proper cross-platform binary generation and enables Windows compatibility.
