+++ pkg/fs/fs.go
@@ -6,7 +6,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"syscall"
 )
 
 //go:generate mockgen -source=fs.go -destination=mockfs.gen.go -package=fs
@@ -237,51 +236,7 @@ func (f *realFS) ExpandPath(path string) (string, error) {
 	return filepath.Join(homeDir, strings.TrimPrefix(path, "~")), nil
 }
 
-// FileLock acquires a file lock and returns an unlock function.
-func (f *realFS) FileLock(filename string) (func(), error) {
-	// Create lock file path
-	lockPath := filename + ".lock"
 
-	// Ensure parent directory exists before creating lock file
-	lockDir := filepath.Dir(lockPath)
-	if err := os.MkdirAll(lockDir, 0755); err != nil {
-		return nil, err
-	}
-
-	// Create lock file
-	lockFile, err := os.Create(lockPath)
-	if err != nil {
-		return nil, err
-	}
-
-	// Acquire file lock (non-blocking)
-	if err := syscall.Flock(int(lockFile.Fd()), syscall.LOCK_EX|syscall.LOCK_NB); err != nil {
-		if closeErr := lockFile.Close(); closeErr != nil {
-			// Log the error but don't fail for cleanup errors
-			_ = closeErr
-		}
-		if removeErr := os.Remove(lockPath); removeErr != nil {
-			// Log the error but don't fail for cleanup errors
-			_ = removeErr
-		}
-		return nil, err
-	}
-
-	// Return unlock function
-	unlock := func() {
-		_ = syscall.Flock(int(lockFile.Fd()), syscall.LOCK_UN)
-		if closeErr := lockFile.Close(); closeErr != nil {
-			// Log the error but don't fail for cleanup errors
-			_ = closeErr
-		}
-		if removeErr := os.Remove(lockPath); removeErr != nil {
-			// Log the error but don't fail for cleanup errors
-			_ = removeErr
-		}
-	}
-
-	return unlock, nil
-}
 
 // CreateFileIfNotExists creates a file with initial content if it doesn't exist.
 func (f *realFS) CreateFileIfNotExists(filename string, initialContent []byte, perm os.FileMode) error {
